Backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==4.2.28
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-filter==23.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
